clean up and remove unused files
