--- components/AllowList/index.tsx
@@ -1,3 +1,0 @@
-import AllowList from "./AllowList"
-
-export default AllowList
