--- components/AllowListForm/index.tsx
@@ -1,3 +1,0 @@
-import AllowListForm from "./AllowListForm"
-
-export default AllowListForm
