--- components/ChoiceScreen/index.tsx
@@ -1,3 +1,0 @@
-import ChoiceScreen from "./ChoiceScreen"
-
-export default ChoiceScreen
