--- components/FindCre8orType/index.tsx
@@ -1,3 +1,0 @@
-import FindCre8orType from "./FindCre8orType"
-
-export default FindCre8orType
