+++ components/GameScreen/GameScreen.tsx
@@ -1,0 +1,3 @@
+const GameScreen = () => <div>GameScreen</div>
+
+export default GameScreen
