+++ components/GameScreen/index.tsx
@@ -1,0 +1,3 @@
+import GameScreen from "./GameScreen"
+
+export default GameScreen
