--- components/PickCre8orType/index.tsx
@@ -1,3 +1,0 @@
-import PickCre8orType from "./pickCre8orType"
-
-export default PickCre8orType
