--- components/Roadmap/index.tsx
@@ -1,3 +1,0 @@
-import Roadmap from "./Roadmap"
-
-export default Roadmap
