--- components/RoadmapPage/index.tsx
@@ -1,3 +1,0 @@
-import RoadmapPage from "./RoadmapPage"
-
-export default RoadmapPage
