+++ pages/_app.tsx
@@ -6,7 +6,7 @@ import { allChains, configureChains, createClient, WagmiConfig } from "wagmi"
 import { publicProvider } from "wagmi/providers/public"
 import { ToastContainer } from "react-toastify"
 import "react-toastify/dist/ReactToastify.css"
-import { Cre8orsProvider } from "../providers/Crea8orsProvider"
+import { MusicGameProvider as MusicGameProvider } from "../providers/MusicGameProvider"
 
 const { chains, provider, webSocketProvider } = configureChains(
   allChains.filter(
@@ -33,10 +33,10 @@ function MyApp({ Component, pageProps }: AppProps) {
   return (
     <WagmiConfig client={wagmiClient}>
       <RainbowKitProvider modalSize="compact" chains={chains}>
-        <Cre8orsProvider>
+        <MusicGameProvider>
           <Component {...pageProps} />
           <ToastContainer />
-        </Cre8orsProvider>
+        </MusicGameProvider>
       </RainbowKitProvider>
     </WagmiConfig>
   )
